# OpenAPI.yaml
openapi: 3.0.0
info:
  title: Machine API
  version: 1.0.0
paths:
  /machine/create:
    post:
      summary: Create a new machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineCreate'
      responses:
        '200':
          description: Machine created successfully
  /machine/get:
    get:
      summary: Get machine
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          description: The id of the machine to retrieve
        - in: query
          name: email
          schema:
            type: string
            format: email
          description: The email of the machine to retrieve
      responses:
        '200':
          description: A list of machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MachineRead'
  /machine/update:
    put:
      summary: Update a machine
      parameters:
        - in: query
          name: machine_id
          schema:
            type: integer
          required: true
          description: The id of the machine to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineUpdate'

      responses:
        '200':
          description: Machine updated successfully
  /machine/schema/{method}:
    get:
      summary: Get the JSON schema for create/update, see below
      parameters:
        - in: path
          name: method
          schema:
            type: string
            enum: [update, create]
          required: true
      responses:
        '200':
          description: The JSON schema of the selected method
          content:
            application/json:
              schema:
                type: object
components:
  schemas:
    MachineBase:
      type: object
      properties:
        name:
          type: string
          maxLength: 10
        location:
          type: string
        email:
          type: string
          format: email
        number:
          type: integer
        float_number:
          type: number
        enum:
          type: string
          enum: [active, not_active]
    MachineCreate:
      allOf:
        - $ref: '#/components/schemas/MachineBase'
        - type: object
          properties:
            password:
              type: string
    MachineUpdate:
      allOf:
        - $ref: '#/components/schemas/MachineBase'
        - type: object
          properties:
            password:
              type: string

    MachineRead:
      allOf:
        - $ref: '#/components/schemas/MachineBase'
        - type: object
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
            edited_at:
              type: string
              format: date-time